Split new-style lineage tests into separate test file.
